requirements.txt
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
binaryornot==0.4.4
certifi==2025.1.31
channels==4.2.0
channels-redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==1.7.3
crispy-tailwind==1.0.3
Django==4.2.20
django-browser-reload==1.18.0
django-crispy-forms==2.3
django-tailwind==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
idna==3.10
jinja2==3.1.6
jinja2-time==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
pillow==11.2.1
poyo==0.5.0
pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
urllib3==1.26.20
